replace placeholder with correct screen
